+++ Runner/src/main/java/xyz/fay/runner/model/MusicQueryResponse.java
@@ -1,13 +1,15 @@
 package xyz.fay.runner.model;
 
+import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
 @Data
 @AllArgsConstructor
+@JsonInclude(JsonInclude.Include.NON_NULL)
 public final class MusicQueryResponse {
-    private final String display;
     private final String ext;
+    private final String fullName;
     private final String name;
     private final String type;
 }