+++ Runner/src/main/java/xyz/fay/runner/utils/StringUtils.java
@@ -1,0 +1,20 @@
+package xyz.fay.runner.utils;
+
+public final class StringUtils {
+    public static String substringBefore(final String string, final String separator) {
+        if (string == null || string.isEmpty() || separator == null) {
+            return string;
+        }
+
+        if (separator.isEmpty()) {
+            return "";
+        }
+
+        final int pos = string.indexOf(separator);
+        if (pos == -1) {
+            return string;
+        }
+
+        return string.substring(0, pos);
+    }
+}